refactor(close): split close command stage modules
Extract the close pipeline into bounded stage owners while preserving the
public facade, shared dependency-injection identity, and lifecycle behavior.

Rebind retention, event, and citation gates to the physical stage modules and
add an anti-reconsolidation structural test.
