+++ docs/observability-event-contract.md
@@ -642,7 +642,7 @@ favor of a plain typed `emit(...)` call, and the kind is added to both
 
 #### close_archive_result
 Category `lifecycle`, severity `notice`, privacy `operational`. Producer
-`src/commands/close.ts:539` (`emitCloseArchiveResult`, called after the clean
+`src/commands/close/archive-stage.ts:38` (`emitCloseArchiveResult`, called after the clean
 stage so `source_disposition` can be finalized to `removed` for cleaned
 artifacts). Consumers: none — owner **#2047**. Retention: **#2047**. The payload
 is one structured result per archived artifact (`requiredness`/`attempt`/
@@ -927,7 +927,7 @@ row 17 records the authoritative knowledge-receipt partition added by #2031.
 | 9 | `.swarm/background-delegations.jsonl` | `src/background/pending-delegations.ts` (checkpoint layer, #2034) | `pr-workflow-session-resolver.ts`, `pr-workflow-gate.ts`, `init-orphan-recovery.ts`, `delegation-gate/worktree-collision-ownership.ts` | `status` | **epoch-ms number** | `schemaVersion` 1\|2\|3 | `correlationId`, `parentSessionId`, `callID`, `jobId`, `planTaskId`, `evidenceTaskId`, `batchId`, `laneId`, `workflowLane`, `worktreeId` | no swarm-run id distinct from `parentSessionId` | bounded — compacted to checkpoint + tail above 1 MiB (#2034) | authoritative | **#2034** |
 | 10 | `.swarm/session/shell-audit.jsonl` | `src/hooks/guardrails/audit-log.ts:496` (validated+redacted decision seam; persistence owned by the bounded store `src/hooks/guardrails/shell-audit-store.ts:786`, issue #2040) | `src/hooks/guardrails/shell-audit-store.ts:787` `readShellAuditTail` / `getShellAuditFoldedSummary` (tail-bounded, manifest-stripped); rendered by `src/services/guardrail-log-service.ts:146`. **Correction:** an earlier draft of this row cited `src/commands/archive.ts` as a reader with invented line numbers and behaviour — that file contains no reference to `shell-audit` at all, and the claim was removed (issue #2029 final-critic B-4). **Post-#2040:** the pre-store whole-file reader no longer exists. | `type` (**stripped for `shell`**) | ISO string (caller-supplied) | `swarm-shell-audit-manifest` v1 header (issue #2040) | `sessionID`, `agent`, `tool` (+ `commandHash` on typed command entries, never rendered) | **no `callID`** → cannot join to row 9 (`background-delegations.jsonl`) | via `session/` dir, finalized as a validated cut first (`close.ts` finalize stage, issue #2040) | operational | **#2040** |
 | 11 | council evidence + `.swarm/council/{taskId}.rounds.jsonl` | `src/council/council-evidence-writer.ts:91` (evidence rewrite at `.swarm/evidence/{taskId}.json`; rounds append) | (council-consuming code paths; not itemized separately from evidence consumers) | none | `synthesis.timestamp` | none (implicit: `quorumSize` defaulted to 1 when absent, `:156-158`) | `sessionId` (= `swarmId`), `roundNumber` | `taskId` only in filename on the rounds log; no `callID` | evidence: yes; `council/`: **no** | authoritative | **#2046** |
-| 12 | `.swarm/archive/swarm-{ts}-{suffix}/` | `src/commands/close.ts:1051-1054` | `src/commands/close.ts:2233-2235` (`readdir` + `startsWith('swarm-')` — finalize-idempotency + reflection filename scans only; nothing re-reads bundle *contents*) | n/a | ISO in path | preserves bytes verbatim | n/a | n/a | is the archive | governed content | **#2030** |
+| 12 | `.swarm/archive/swarm-{ts}-{suffix}/` | `src/commands/close/archive-stage.ts:146-153` | `src/commands/close/orchestrator.ts:142-146` (`readdir` + `startsWith('swarm-')` — finalize-idempotency + reflection filename scans only; nothing re-reads bundle *contents*) | n/a | ISO in path | preserves bytes verbatim | n/a | n/a | is the archive | governed content | **#2030** |
 | 13 | SQLite `memory_events` | `src/memory/sqlite-provider.ts:200-207` | memory-provider internal readers (not itemized here — index design is #2048's scope) | `operation` column | `timestamp` column, ISO string | table has no explicit version column (SQLite `_meta` table tracks migration version 4+, not per-row) | `target_id` | no `session_id`/`task_id` column | **no** | authoritative | **#2036** (retention), #2048 (index) |
 | 14 | SQLite `memory_recall_usage` | `src/memory/sqlite-provider.ts:209-214` | memory-provider internal readers | `bundle_id` column | `timestamp` column, ISO string | none (migration-versioned schema, not row-versioned) | `bundle_id`, `run_id` (added migration v9) | no `session_id`/`task_id` column | **no** | derived | **#2036**, #2048 |
 | 15 | SQLite `memory_reward_events` | `src/memory/sqlite-provider.ts:286` | memory-provider internal readers | `verdict` column | `timestamp` column, ISO string | none | `memory_id`, `run_id`, `unit_id` | no `session_id`/`task_id` column | **no** | derived | **#2036**, #2048 |
