+++ scripts/retention-registry.data.ts
@@ -207,7 +207,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		},
 		lockModel: 'none — single module-scoped write stream; rotation only from the throttled emit path',
 		crashBehavior: 'append stream; torn trailing line tolerated (both readers JSON.parse try/catch); flushAndDrainTelemetry replaces stream handle before end() (src/telemetry.ts:438-469)',
-		closePolicy: 'archived+cleaned — flush (close.ts:1274), ARCHIVE_ARTIFACTS (close.ts:400-401), ACTIVE_STATE_TO_CLEAN (close.ts:490-495)',
+		closePolicy: 'archived+cleaned — flush (src/commands/close/archive-stage.ts:161-204), ARCHIVE_ARTIFACTS (src/commands/close/constants.ts:50-51), ACTIVE_STATE_TO_CLEAN (src/commands/close/constants.ts:169-174)',
 		closeArrayMembership: {
 			'telemetry.jsonl': 'archive+clean',
 			'telemetry.jsonl.1': 'archive+clean',
@@ -509,7 +509,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		lockModel: 'withEvidenceLock agent=background on every mutation (:170-173); reads lock-free',
 		crashBehavior:
 			'torn append tolerated by lenient fold, strict recovery fails closed; manifest-gated checkpoint publication — checkpoint without manifest ignored (:1614-1626)',
-		closePolicy: 'archived-only — ARCHIVE_ARTIFACTS (close.ts:463-465); deliberately NOT cleaned (cross-session store; compaction is the bounded-retention mechanism, close.ts:458-466 docblock)',
+		closePolicy: 'archived-only — ARCHIVE_ARTIFACTS (src/commands/close/constants.ts:75-77); deliberately NOT cleaned (cross-session store; compaction is the bounded-retention mechanism, src/commands/close/constants.ts:70-78 docblock)',
 		closeArrayMembership: {
 			'background-delegations.jsonl': 'archive-only',
 			'background-delegations.checkpoint.json': 'archive-only',
@@ -540,7 +540,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		readBound: { pattern: 'indexed', bound: 'single small JSON artifact', sync: true, citation: 'src/background/delegation-health.ts:309' },
 		lockModel: 'written under the store lock (compaction/recovery) or best-effort',
 		crashBehavior: 'atomic rename with 5×15 ms Windows retry; best-effort, returns null on failure (:419)',
-		closePolicy: 'archived-only — ARCHIVE_ARTIFACTS (close.ts:420), not cleaned',
+		closePolicy: 'archived-only — ARCHIVE_ARTIFACTS (src/commands/close/constants.ts:78), not cleaned',
 		closeArrayMembership: {
 			'background-delegations-health.json': 'archive-only',
 		},
@@ -676,7 +676,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 			bound: 'MAX_PR_FEEDBACK_MONITOR_EVENTS 20 per queue (:14); MAX_QUEUE_BYTES 512 KiB per file (:15); in-memory cache MAX_TRACKED_SESSIONS 200 FIFO (:16); the retention sweep\'s pr-feedback-events family age-prunes every queue file at 30 d (src/retention/sweep.ts:91)',
 			scope: 'per-key',
 			keyspaceBound:
-				'FINITE BY REAPER: the retention-sweep family pr-feedback-events (src/retention/sweep.ts:91) age-prunes every session queue file at 30 d — wired post-init (src/index.ts:1293) and pre-close (src/commands/close.ts:2626) — so the session-file keyspace cannot outgrow the sweep horizon. The in-process MAX_TRACKED_SESSIONS=200 FIFO (src/background/pr-feedback-event-queue.ts:16) remains an in-memory bound only. This closes the #2038-class keyspace gap this row recorded under #2309.',
+				'FINITE BY REAPER: the retention-sweep family pr-feedback-events (src/retention/sweep.ts:91) age-prunes every session queue file at 30 d — wired post-init (src/index.ts:1293) and pre-close (src/commands/close/orchestrator.ts:269-286) — so the session-file keyspace cannot outgrow the sweep horizon. The in-process MAX_TRACKED_SESSIONS=200 FIFO (src/background/pr-feedback-event-queue.ts:16) remains an in-memory bound only. This closes the #2038-class keyspace gap this row recorded under #2309.',
 			citation: 'src/background/pr-feedback-event-queue.ts:14-19; src/retention/sweep.ts:91',
 		},
 		readBound: { pattern: 'indexed', bound: '≤512 KiB hard read bound', sync: false, citation: 'src/background/pr-feedback-event-queue.ts:440-470' },
@@ -690,7 +690,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		disposition: {
 			kind: 'not-a-defect',
 			proof:
-				'The #2483 retention sweep\'s pr-feedback-events family age-prunes every session queue file at 30 d (src/retention/sweep.ts:91), wired post-init (src/index.ts:1293) and pre-close (src/commands/close.ts:2626) — the keyspace gap behind the #2309/#2038 reclassification is closed; per-file caps 20 events / 512 KiB (src/background/pr-feedback-event-queue.ts:14-15) bound each key.',
+				'The #2483 retention sweep\'s pr-feedback-events family age-prunes every session queue file at 30 d (src/retention/sweep.ts:91), wired post-init (src/index.ts:1293) and pre-close (src/commands/close/orchestrator.ts:269-286) — the keyspace gap behind the #2309/#2038 reclassification is closed; per-file caps 20 events / 512 KiB (src/background/pr-feedback-event-queue.ts:14-15) bound each key.',
 		},
 	},
 	{
@@ -951,12 +951,12 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		readBound: { pattern: 'directory-scan', bound: 'live locks only (expired filtered)', sync: true, citation: 'src/parallel/file-locks.ts:303-369' },
 		lockModel: 'IS the lock infrastructure (proper-lockfile + sentinels)',
 		crashBehavior: 'proper-lockfile stale recovery; orphaned .meta falls back to mtime expiry (:344-369)',
-		closePolicy: 'untouched — deliberately excluded from close (close.ts:1480-1481: "per-run locks are managed via proper-lockfile, not archived or cleaned by close")',
+		closePolicy: 'untouched — deliberately excluded from close (src/commands/close/archive-stage.ts:423-424: "per-run locks are managed via proper-lockfile, not archived or cleaned by close")',
 		resetPolicy: 'not reset',
 		legacyCompatibility: 'legacy base64 lock paths honored (:82-94)',
 		healthSignal: 'listActiveLocks output',
 		owner: '#2035 (merged)',
-		disposition: { kind: 'not-a-defect', proof: 'Per-run lock machinery with stale-timeout cleanup and deliberate close exclusion (src/parallel/file-locks.ts:7,250-297; close.ts:1480-1481).' },
+		disposition: { kind: 'not-a-defect', proof: 'Per-run lock machinery with stale-timeout cleanup and deliberate close exclusion (src/parallel/file-locks.ts:7,250-297; src/commands/close/archive-stage.ts:423-424).' },
 	},
 	{
 		id: 'task-evidence-trajectory',
@@ -978,13 +978,13 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 			bound: 'max_lines PER-FILE enforced at write time; truncation keeps newest floor(max_lines/2) (:114-133). NOTE (maintainer review #2395, finding on claim 7): the production knob is now `prm.max_trajectory_lines` (src/index.ts passes it to createTrajectoryLoggerHook and the denied-call path), previously a hardcoded 1000 — the writer CODE is unchanged but the evidence truncation budget is config-coupled; default 1000 is identical, so no default-config regression.',
 			scope: 'per-key',
 			keyspaceBound:
-				'FINITE BY REAPER, not by key domain: one key per taskId directory, and taskId is only shape-validated (src/validation/task-id.ts:69-114 admits any /^[a-zA-Z0-9][a-zA-Z0-9._-]*$/ id, explicitly beyond plan task numbers), so the key domain is open. What bounds it is a GLOBAL deleter: ACTIVE_STATE_DIRS_TO_CLEAN contains "evidence" (src/commands/close.ts:574-583) and the close clean loop runs fs.rm(dirPath, {recursive: true}) over the whole .swarm/evidence/ tree (src/commands/close.ts:1761-1773), dropping every key in one pass on a single session-lifecycle trigger. CAVEAT (verified, do not soften): the trigger is lifecycle-driven rather than size-driven and is archive-first-gated (src/commands/close.ts:1762-1765 skips the delete when the directory was not archived), and neither /swarm reset nor /swarm reset-session touches this tree — so a session that never closes accumulates one directory per distinct taskId.',
+				'FINITE BY REAPER, not by key domain: one key per taskId directory, and taskId is only shape-validated (src/validation/task-id.ts:69-114 admits any /^[a-zA-Z0-9][a-zA-Z0-9._-]*$/ id, explicitly beyond plan task numbers), so the key domain is open. What bounds it is a GLOBAL deleter: ACTIVE_STATE_DIRS_TO_CLEAN contains "evidence" (src/commands/close/constants.ts:253-269) and the close clean loop runs fs.rm(dirPath, {recursive: true}) over the whole .swarm/evidence/ tree (src/commands/close/clean-stage.ts:176-190), dropping every key in one pass on a single session-lifecycle trigger. CAVEAT (verified, do not soften): the trigger is lifecycle-driven rather than size-driven and is archive-first-gated (src/commands/close/clean-stage.ts:176-185 skips the delete when the directory was not archived), and neither /swarm reset nor /swarm reset-session touches this tree — so a session that never closes accumulates one directory per distinct taskId.',
 			citation: 'src/hooks/trajectory-logger.ts:114-133,466,558',
 		},
 		readBound: { pattern: 'full-file', bound: '≤500 lines per file by write-side truncation', sync: false, citation: 'src/hooks/trajectory-logger.ts:558 + micro-reflector.ts:262' },
 		lockModel: 'none — single-line appends; per-line JSONL validity preserved',
 		crashBehavior: 'torn tail tolerated by reader; truncation runs after append so newest entry is durable first',
-		closePolicy: 'cleaned — evidence/ dir archived+cleaned (ACTIVE_STATE_DIRS_TO_CLEAN close.ts:568)',
+		closePolicy: 'cleaned — evidence/ dir archived+cleaned (ACTIVE_STATE_DIRS_TO_CLEAN src/commands/close/constants.ts:256)',
 		resetPolicy: 'reset-session does not touch evidence/',
 		legacyCompatibility: 'reader skips malformed lines',
 		healthSignal: 'n/a',
@@ -1160,13 +1160,13 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 			bound: 'one bounded-input report per plan (inputs capped :38-43); idempotent dedup',
 			scope: 'per-key',
 			keyspaceBound:
-				'FINITE BY REAPER: one key per planId (.swarm/post-mortem-{planId}.md). The GLOBAL deleter is a directory scan, not a per-plan hook — /swarm close readdirs .swarm/ and matches every /^post-mortem-[^/\\\\]+\\.md$/ name (src/commands/close.ts:1447-1452), copies each into the archive, records it in ctx.archivedActiveStateFiles, and then unlinks each matched artifact from .swarm/ (src/commands/close.ts:1738-1747). Because the sweep is driven by the directory listing rather than by a known set of plan ids, it reclaims keys the writer never told it about. CAVEAT: the unlink runs only for artifacts that archived successfully — src/commands/close.ts:1732-1736 deliberately preserves active state when nothing was archived — and the trigger is session close, so an unclosed session retains one report per plan.',
+				'FINITE BY REAPER: one key per planId (.swarm/post-mortem-{planId}.md). The GLOBAL deleter is a directory scan, not a per-plan hook — /swarm close readdirs .swarm/ and matches every /^post-mortem-[^/\\\\]+\\.md$/ name (src/commands/close/archive-stage.ts:351-375), copies each into the archive, records it in ctx.archivedActiveStateFiles, and then unlinks each matched artifact from .swarm/ (src/commands/close/clean-stage.ts:155-174). Because the sweep is driven by the directory listing rather than by a known set of plan ids, it reclaims keys the writer never told it about. CAVEAT: the unlink runs only for artifacts that archived successfully — src/commands/close/clean-stage.ts:45-79 deliberately preserves active state when nothing was archived — and the trigger is session close, so an unclosed session retains one report per plan.',
 			citation: 'src/hooks/curator-postmortem.ts:38-43,1129-1137',
 		},
 		readBound: { pattern: 'full-file', bound: 'single report per plan', sync: true, citation: 'src/hooks/curator-postmortem.ts:819' },
 		lockModel: 'tryAcquireLock non-blocking, skip if held (:836-846,1145-1156)',
 		crashBehavior: 'atomic write; invalid report regenerated next run',
-		closePolicy: 'archived+cleaned — dynamic artifacts (close.ts:1418-1477,1709-1728)',
+		closePolicy: 'archived+cleaned — dynamic artifacts (src/commands/close/archive-stage.ts:351-375; src/commands/close/clean-stage.ts:155-174)',
 		resetPolicy: 'not reset',
 		legacyCompatibility: 'parseLegacyPostMortemActions + structured parser both tried (:249,:400,:475-490)',
 		healthSignal: 'n/a',
@@ -1184,7 +1184,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		schemaVersion: 'none',
 		stateClass: 'operational',
 		privacyClass: 'metadata',
-		writeLimits: { bound: 'no per-file cap; bounded by session — evidence/ dir archived+cleaned at close (close.ts:568)', scope: 'session-scoped', citation: 'src/turbo/epic/promotion-evidence.ts:61; close.ts:565-574' },
+		writeLimits: { bound: 'no per-file cap; bounded by session — evidence/ dir archived+cleaned at close (src/commands/close/constants.ts:256)', scope: 'session-scoped', citation: 'src/turbo/epic/promotion-evidence.ts:61; src/commands/close/constants.ts:253-269' },
 		readBound: { pattern: 'full-file', bound: 'session-scoped file (close-cleaned)', sync: true, citation: 'src/turbo/epic/promotion-evidence.ts:70-88' },
 		lockModel: 'none — single-line appends',
 		crashBehavior: 'torn tail skipped by reader',
@@ -1253,14 +1253,14 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		stateClass: 'operational',
 		privacyClass: 'metadata',
 		writeLimits: {
-			bound: 'MAX_CALIBRATION_MODULES 500 truncation on every save AND load (src/turbo/epic/calibration.ts:70,74-79); MAX_DIVERGENCE_BYTES 8 MiB write-side compaction with whole-record floor (src/turbo/epic/divergence-recorder.ts:64,76-77); retention sweep deletes epic/calibration.json + epic/divergence.jsonl whole-file at 30 d (src/retention/sweep.ts:169); close archives+cleans epic-state.json + turbo-state.json (close.ts:441-442,570-571) and the runs/ + epic/ dirs (close.ts:648-649)',
+			bound: 'MAX_CALIBRATION_MODULES 500 truncation on every save AND load (src/turbo/epic/calibration.ts:70,74-79); MAX_DIVERGENCE_BYTES 8 MiB write-side compaction with whole-record floor (src/turbo/epic/divergence-recorder.ts:64,76-77); retention sweep deletes epic/calibration.json + epic/divergence.jsonl whole-file at 30 d (src/retention/sweep.ts:169); close archives+cleans epic-state.json + turbo-state.json (src/commands/close/constants.ts:53-54; src/commands/close/constants.ts:187-190) and the runs/ + epic/ dirs (src/commands/close/constants.ts:267-268)',
 			scope: 'global',
-			citation: 'src/turbo/epic/calibration.ts:70; src/turbo/epic/divergence-recorder.ts:64; src/retention/sweep.ts:169; src/commands/close.ts:570',
+			citation: 'src/turbo/epic/calibration.ts:70; src/turbo/epic/divergence-recorder.ts:64; src/retention/sweep.ts:169; src/commands/close/constants.ts:253-269',
 		},
 		readBound: { pattern: 'mixed full-file + tail', bound: 'divergence reader tail-bounded 16 MiB (MAX_TAIL_BYTES); calibration reader full-file but the state is cap-truncated on load', sync: true, citation: 'src/turbo/epic/divergence-recorder.ts:247,264' },
 		lockModel: 'SQLite BEGIN IMMEDIATE + revision/generation CAS for session state; recovery records cleared on successful merge-back',
 		crashBehavior: 'atomic rewrites with fail-closed corrupt markers and repair paths',
-		closePolicy: 'archived+cleaned — epic-state.json + turbo-state.json in both arrays (close.ts:441-442,570-571); runs/ + epic/ dirs archived+cleaned (ACTIVE_STATE_DIRS_TO_CLEAN close.ts:648-649); recovery/ stays out of close — the 30 d sweep owns it (src/retention/sweep.ts:102)',
+		closePolicy: 'archived+cleaned — epic-state.json + turbo-state.json in both arrays (src/commands/close/constants.ts:53-54; src/commands/close/constants.ts:187-190); runs/ + epic/ dirs archived+cleaned (ACTIVE_STATE_DIRS_TO_CLEAN src/commands/close/constants.ts:267-268); recovery/ stays out of close — the 30 d sweep owns it (src/retention/sweep.ts:102)',
 		closeArrayMembership: {
 			'epic-state.json': 'archive+clean',
 			'turbo-state.json': 'archive+clean',
@@ -1272,7 +1272,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		disposition: {
 			kind: 'not-a-defect',
 			proof:
-				'Every member is bounded: calibration MAX_CALIBRATION_MODULES 500 truncation on save AND load (src/turbo/epic/calibration.ts:70,74-79), divergence MAX_DIVERGENCE_BYTES 8 MiB write-side compaction (src/turbo/epic/divergence-recorder.ts:64,76-77), the retention sweep\'s 30 d whole-file delete for both epic diagnostics (src/retention/sweep.ts:169), and the #2483 close wiring (epic-state.json + turbo-state.json archive+clean, close.ts:441-442,570-571; runs/ + epic/ dirs, close.ts:648-649).',
+				'Every member is bounded: calibration MAX_CALIBRATION_MODULES 500 truncation on save AND load (src/turbo/epic/calibration.ts:70,74-79), divergence MAX_DIVERGENCE_BYTES 8 MiB write-side compaction (src/turbo/epic/divergence-recorder.ts:64,76-77), the retention sweep\'s 30 d whole-file delete for both epic diagnostics (src/retention/sweep.ts:169), and the #2483 close wiring (epic-state.json + turbo-state.json archive+clean, src/commands/close/constants.ts:53-54; src/commands/close/constants.ts:187-190; runs/ + epic/ dirs, src/commands/close/constants.ts:267-268).',
 		},
 	},
 	{
@@ -1289,7 +1289,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		schemaVersion: 'none',
 		stateClass: 'governed-content',
 		privacyClass: 'content',
-		writeLimits: { bound: 'per-phase/per-lane artifacts; evidence/ dir archived+cleaned at close', scope: 'session-scoped', citation: 'close.ts:565-574' },
+		writeLimits: { bound: 'per-phase/per-lane artifacts; evidence/ dir archived+cleaned at close', scope: 'session-scoped', citation: 'src/commands/close/constants.ts:253-269' },
 		readBound: { pattern: 'indexed', bound: 'session-scoped evidence dir', sync: true, citation: 'src/turbo/lean/evidence.ts:234-341' },
 		lockModel: 'none explicit; atomic temp+rename per file with cache invalidation',
 		crashBehavior: 'atomic; ENOENT → null',
@@ -1326,7 +1326,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		schemaVersion: 'per-artifact Zod schemas (bounded fields, e.g. findings ≤4000 chars)',
 		stateClass: 'governed-content',
 		privacyClass: 'mixed',
-		writeLimits: { bound: 'per-phase overwrite or bounded per-run artifacts; whole tree archived+cleaned at close', scope: 'session-scoped', citation: 'close.ts:565-574; write-hallucination-evidence.ts:34' },
+		writeLimits: { bound: 'per-phase overwrite or bounded per-run artifacts; whole tree archived+cleaned at close', scope: 'session-scoped', citation: 'src/commands/close/constants.ts:253-269; write-hallucination-evidence.ts:34' },
 		readBound: { pattern: 'indexed', bound: 'single small JSON per gate', sync: true, citation: 'src/tools/phase-complete/gates/mutation-gate.ts' },
 		lockModel: 'atomic temp+rename (no lock — single-writer tools)',
 		crashBehavior: 'atomic; old file preserved on failed rename',
@@ -1399,7 +1399,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		schemaVersion: 'none',
 		stateClass: 'governed-content',
 		privacyClass: 'metadata',
-		writeLimits: { bound: 'one line per attestation decision; evidence/ dir archived+cleaned at close', scope: 'session-scoped', citation: 'close.ts:565-574' },
+		writeLimits: { bound: 'one line per attestation decision; evidence/ dir archived+cleaned at close', scope: 'session-scoped', citation: 'src/commands/close/constants.ts:253-269' },
 		readBound: { pattern: 'write-only', bound: 'n/a', sync: false, citation: 'no production reader (verified)' },
 		lockModel: 'none',
 		crashBehavior: 'append propagates errors to caller; torn tail unobserved (no reader)',
@@ -1438,12 +1438,12 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		readBound: { pattern: 'directory-scan', bound: '≤10k files scanned, ≤2 MiB per scope read', sync: true, citation: 'src/scope/scope-persistence.ts:1345-1411,2534' },
 		lockModel: 'proper-lockfile per binding with generation-lock sidecars (:347,:848-852); bounded sync retry (:112-129)',
 		crashBehavior: 'atomic writes; v1 write silent-no-op, v2 returns typed failures with rollback (:902-913)',
-		closePolicy: 'cleaned-only — clearAllScopes rmSync, NOT archived ("scope files are ephemeral state", close.ts:1912-1916)',
+		closePolicy: 'cleaned-only — clearAllScopes rmSync, NOT archived ("scope files are ephemeral state", src/commands/close/clean-stage.ts:358-362)',
 		resetPolicy: 'reset-session does not clear scopes (close does)',
 		legacyCompatibility: 'v1 disk + plan.json + pending-map fallback chain (:2655)',
 		healthSignal: 'typed persistence results',
 		owner: 'this-gate',
-		disposition: { kind: 'not-a-defect', proof: 'Fully bounded constant set (10k/2 MiB/256/7d) with per-binding locks and deliberate ephemeral close cleanup (src/scope/scope-persistence.ts:87-92; close.ts:1912-1916).' },
+		disposition: { kind: 'not-a-defect', proof: 'Fully bounded constant set (10k/2 MiB/256/7d) with per-binding locks and deliberate ephemeral close cleanup (src/scope/scope-persistence.ts:87-92; src/commands/close/clean-stage.ts:358-362).' },
 	},
 	{
 		id: 'task-workflow-evidence',
@@ -1470,8 +1470,8 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 			bound: 'retryHistory ≤3 (schema :303); per-task file; evidence/ archived+cleaned at close',
 			scope: 'per-key',
 			keyspaceBound:
-				'FINITE BY REAPER, not by key domain: one key per taskId — a flat .swarm/evidence/{taskId}.json (src/gate-evidence.ts:764 getEvidencePath) whose taskId is only shape-validated (src/validation/task-id.ts:69-114), so the domain is open. The GLOBAL deleter is the same one the task-evidence-trajectory row cites: "evidence" is in ACTIVE_STATE_DIRS_TO_CLEAN (src/commands/close.ts:574-583) and the close clean loop recursively removes the whole tree (src/commands/close.ts:1761-1773), taking every {taskId}.json with it. Note the per-file retryHistory ≤3 cap is NOT the keyspace bound — it caps one key\'s history and says nothing about how many keys exist. CAVEAT: archive-first-gated (src/commands/close.ts:1762-1765) and untouched by /swarm reset and /swarm reset-session, so an unclosed session holds one file per distinct taskId.',
-			citation: 'src/gate-evidence.ts:303; src/commands/close.ts:574 ACTIVE_STATE_DIRS_TO_CLEAN',
+				'FINITE BY REAPER, not by key domain: one key per taskId — a flat .swarm/evidence/{taskId}.json (src/gate-evidence.ts:764 getEvidencePath) whose taskId is only shape-validated (src/validation/task-id.ts:69-114), so the domain is open. The GLOBAL deleter is the same one the task-evidence-trajectory row cites: "evidence" is in ACTIVE_STATE_DIRS_TO_CLEAN (src/commands/close/constants.ts:253-269) and the close clean loop recursively removes the whole tree (src/commands/close/clean-stage.ts:176-190), taking every {taskId}.json with it. Note the per-file retryHistory ≤3 cap is NOT the keyspace bound — it caps one key\'s history and says nothing about how many keys exist. CAVEAT: archive-first-gated (src/commands/close/clean-stage.ts:176-185) and untouched by /swarm reset and /swarm reset-session, so an unclosed session holds one file per distinct taskId.',
+			citation: 'src/gate-evidence.ts:303; src/commands/close/constants.ts:253-269 ACTIVE_STATE_DIRS_TO_CLEAN',
 		},
 		readBound: { pattern: 'full-file', bound: 'single per-task JSON', sync: true, citation: 'src/gate-evidence.ts:1054-1089' },
 		lockModel: 'withTaskEvidenceLock (evidence/{taskId}.json key) — proper-lockfile, 60 s timeout, backoff+jitter',
@@ -1610,7 +1610,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 			bound: 'MAX_TASK_GATE_REQUIREMENTS_BYTES 256 KiB per task file (:13, hard-fails OVERSIZED); repaired evidence MAX_TASK_GATE_EVIDENCE_BYTES 256 KiB (:33); quarantine caps 32 files / 12 MiB / 128 entries / 768 KiB per record (:34-37)',
 			scope: 'per-key',
 			keyspaceBound:
-				'Two sub-streams, each finite for a different reason. (1) Requirements files, one key per taskId at .swarm/evidence/task-gate-requirements/{taskId}.jsonl (src/evidence/task-gate-requirements.ts:44) — a true SUBPATH of .swarm/evidence/, so the close-time recursive removal of the whole evidence tree (ACTIVE_STATE_DIRS_TO_CLEAN contains "evidence", src/commands/close.ts:574-583; fs.rm recursive at src/commands/close.ts:1761-1773) is a global deleter over this keyspace too. (2) Quarantine sidecars carry a genuine GLOBAL count cap independent of the close lifecycle: TASK_GATE_EVIDENCE_QUARANTINE_DIR is a single shared directory with NO {taskId} path component (src/evidence/task-gate-repair.ts:29-32), and the admission check opendirs it and sums count/bytes across ALL tasks before admitting a write (src/evidence/task-gate-repair.ts:448-480), so total quarantine files can never exceed 32 / 12 MiB however many taskIds appear. CAVEAT: cap (2) bounds by REFUSAL, not eviction — at the ceiling it throws TASK_GATE_EVIDENCE_QUARANTINE_FULL and admits nothing new rather than making room; and sub-stream (1) inherits the evidence-tree caveat (archive-first-gated, untouched by /swarm reset and /swarm reset-session).',
+				'Two sub-streams, each finite for a different reason. (1) Requirements files, one key per taskId at .swarm/evidence/task-gate-requirements/{taskId}.jsonl (src/evidence/task-gate-requirements.ts:44) — a true SUBPATH of .swarm/evidence/, so the close-time recursive removal of the whole evidence tree (ACTIVE_STATE_DIRS_TO_CLEAN contains "evidence", src/commands/close/constants.ts:253-269; fs.rm recursive at src/commands/close/clean-stage.ts:176-190) is a global deleter over this keyspace too. (2) Quarantine sidecars carry a genuine GLOBAL count cap independent of the close lifecycle: TASK_GATE_EVIDENCE_QUARANTINE_DIR is a single shared directory with NO {taskId} path component (src/evidence/task-gate-repair.ts:29-32), and the admission check opendirs it and sums count/bytes across ALL tasks before admitting a write (src/evidence/task-gate-repair.ts:448-480), so total quarantine files can never exceed 32 / 12 MiB however many taskIds appear. CAVEAT: cap (2) bounds by REFUSAL, not eviction — at the ceiling it throws TASK_GATE_EVIDENCE_QUARANTINE_FULL and admits nothing new rather than making room; and sub-stream (1) inherits the evidence-tree caveat (archive-first-gated, untouched by /swarm reset and /swarm reset-session).',
 			citation: 'src/evidence/task-gate-requirements.ts:13,149-175; src/evidence/task-gate-repair.ts:33-37',
 		},
 		readBound: { pattern: 'line-bounded', bound: 'reads reject files over 256 KiB (typed error)', sync: false, citation: 'src/evidence/task-gate-requirements.ts:149-175' },
@@ -1643,7 +1643,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 			bound: 'MAX_BASELINE_FINDINGS 2000 (:65); MAX_BASELINE_BYTES 2 MiB (:68) with truncation :690-712,777-795',
 			scope: 'per-key',
 			keyspaceBound:
-				'FINITE BY REAPER ONLY — the key domain is open. `phase` is a positive integer with no upper limit: the tool schema is z.number().int().min(1) with no .max() (src/tools/sast-scan.ts:969-976, duplicated at src/tools/pre-check-batch.ts:1677-1684), validatePhase rejects only phase < 1 (src/tools/sast-baseline.ts:447-452), and PlanSchema.phases is z.array(PhaseSchema).min(1) with no cap on phase count (src/config/plan-schema.ts:110) — there is no closed phase enum anywhere on this path. What bounds it is the close-time evidence wipe: baselineRelPath resolves through validateSwarmPath (src/tools/sast-baseline.ts:191,512) to .swarm/evidence/{phase}/sast-baseline.json, a true subpath of the tree that ACTIVE_STATE_DIRS_TO_CLEAN removes recursively at /swarm close (src/commands/close.ts:574-583,1761-1773). CAVEAT (verified): the taskId-keyed archiveEvidence retention does NOT reach these directories — it looks for evidence/{taskId}/evidence.json and silently skips a phase directory that has none (src/evidence/manager.ts:676-738) — so /swarm close is the ONLY reclaim path here, and one directory accumulates per distinct phase number until it runs.',
+				'FINITE BY REAPER ONLY — the key domain is open. `phase` is a positive integer with no upper limit: the tool schema is z.number().int().min(1) with no .max() (src/tools/sast-scan.ts:969-976, duplicated at src/tools/pre-check-batch.ts:1677-1684), validatePhase rejects only phase < 1 (src/tools/sast-baseline.ts:447-452), and PlanSchema.phases is z.array(PhaseSchema).min(1) with no cap on phase count (src/config/plan-schema.ts:110) — there is no closed phase enum anywhere on this path. What bounds it is the close-time evidence wipe: baselineRelPath resolves through validateSwarmPath (src/tools/sast-baseline.ts:191,512) to .swarm/evidence/{phase}/sast-baseline.json, a true subpath of the tree that ACTIVE_STATE_DIRS_TO_CLEAN removes recursively at /swarm close (src/commands/close/constants.ts:253-269; src/commands/close/clean-stage.ts:176-190). CAVEAT (verified): the taskId-keyed archiveEvidence retention does NOT reach these directories — it looks for evidence/{taskId}/evidence.json and silently skips a phase directory that has none (src/evidence/manager.ts:676-738) — so /swarm close is the ONLY reclaim path here, and one directory accumulates per distinct phase number until it runs.',
 			citation: 'src/tools/sast-baseline.ts:65-68',
 		},
 		readBound: { pattern: 'full-file', bound: '≤2 MiB by write-side cap', sync: true, citation: 'src/tools/sast-baseline.ts:68,861' },
@@ -1686,10 +1686,10 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 			scope: 'global',
 			citation: 'docs/plan-durability.md; src/plan/ledger.ts:949,1268; src/plan/manager.ts:1467-1491',
 		},
-		readBound: { pattern: 'full-file', bound: 'authoritative-lifecycle rationale: the ledger IS the plan authority (invariant 5); replay is the correctness path, close archives + removes terminal state', sync: true, citation: 'src/plan/ledger.ts:1756-1917; close.ts:1893-1910' },
+		readBound: { pattern: 'full-file', bound: 'authoritative-lifecycle rationale: the ledger IS the plan authority (invariant 5); replay is the correctness path, close archives + removes terminal state', sync: true, citation: 'src/plan/ledger.ts:1756-1917; src/commands/close/clean-stage.ts:324-356' },
 		lockModel: 'withEvidenceLock on the ledger path + optimistic CAS retry (appendLedgerEventWithRetry :1092-1133)',
 		crashBehavior: 'fsync-then-rename append; torn tail quarantined to a side file, prefix-only projection refused (M1 fix)',
-		closePolicy: 'archived + terminal-state REMOVED unconditionally so a closed plan cannot resurrect (close.ts:1893-1910); ledger siblings removed (close.ts:1775-1795)',
+		closePolicy: 'archived + terminal-state REMOVED unconditionally so a closed plan cannot resurrect (src/commands/close/clean-stage.ts:324-356); ledger siblings removed (src/commands/close/clean-stage.ts:221-241)',
 		closeArrayMembership: {
 			'plan-ledger.jsonl': 'archive+clean',
 		},
@@ -1747,12 +1747,12 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		readBound: { pattern: 'full-file', bound: 'single checkpoint document / ≤20-entry log', sync: true, citation: 'src/tools/checkpoint.ts:175; src/plan/checkpoint.ts:90-119' },
 		lockModel: 'withCheckpointMutationLock (proper-lockfile on checkpoints.json, tools/checkpoint.ts:545-644)',
 		crashBehavior: 'temp+rename; git commit precedes checkpoint recording',
-		closePolicy: 'archived + removed from all three locations (close.ts:1808-1833)',
+		closePolicy: 'archived + removed from all three locations (src/commands/close/clean-stage.ts:254-286)',
 		resetPolicy: 'close removes',
 		legacyCompatibility: 'three-location reads with deprecation warnings; cleanup removes all three',
 		healthSignal: 'retention events (checkpoint_retention_applied)',
 		owner: 'this-gate',
-		disposition: { kind: 'not-a-defect', proof: 'Scoped export artifacts (issue #852) with FIFO-bounded checkpoint log and three-location close cleanup (src/tools/checkpoint.ts:217; close.ts:1808-1833).' },
+		disposition: { kind: 'not-a-defect', proof: 'Scoped export artifacts (issue #852) with FIFO-bounded checkpoint log and three-location close cleanup (src/tools/checkpoint.ts:217; src/commands/close/clean-stage.ts:254-286).' },
 	},
 	{
 		id: 'evidence-bundles',
@@ -1762,7 +1762,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		writerModules: ['src/evidence/manager.ts', 'src/tools/write-retro.ts', 'src/tools/sbom-generate.ts', 'src/summaries/manager.ts'],
 		writerCitations: [
 			'src/evidence/manager.ts:193 saveEvidence — withEvidenceLock, bundle trim MAX_BUNDLE_ENTRIES 100 (:261-265), EVINENCE_MAX_JSON_BYTES 500 KiB, temp+rename :286-296',
-			'src/evidence/manager.ts:665 archiveEvidence — retention maxAgeDays 30 / maxBundles 10 defaults (close.ts:1560-1562)',
+			'src/evidence/manager.ts:665 archiveEvidence — retention maxAgeDays 30 / maxBundles 10 defaults (src/commands/close/archive-stage.ts:498-544)',
 			'src/summaries/manager.ts:132 storeSummary — atomic (summaries row below owns lifecycle)',
 		],
 		readerCitations: ['src/evidence/manager.ts:387 loadEvidence — full-file Zod-validated with legacy migration, async; :515 listEvidenceTaskIds — directory list'],
@@ -1773,8 +1773,8 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 			bound: '≤100 entries + ≤500 KiB per bundle; retention 30 d / 10 bundles; evidence/ close-scoped',
 			scope: 'per-key',
 			keyspaceBound:
-				'FINITE BY A REAL GLOBAL REAPER: one key per taskId bundle directory, and archiveEvidence enumerates the ENTIRE keyspace via listEvidenceTaskIds(directory) and deletes across it (src/evidence/manager.ts:676-738) — the maxAgeDays/maxBundles retention selects victims by comparing every remaining bundle against a global count (src/evidence/manager.ts:719-730), so unlike the ≤100-entry/≤500-KiB per-bundle cap in this row\'s `bound` field, that retention IS a keyspace bound. Belt-and-braces, "evidence" is also in ACTIVE_STATE_DIRS_TO_CLEAN and the whole tree is recursively removed at close (src/commands/close.ts:574-583,1761-1773). CAVEAT: archiveEvidence has exactly two callers — /swarm close (src/commands/close.ts:1622, 30 d / 10 bundles), where the subsequent full-tree removal makes it largely redundant, and the operator-invoked /swarm archive (src/commands/archive.ts:95, 90 d / 1000 bundles). It is not wired to any automatic interval, so mid-session reclamation requires one of those two commands to run.',
-			citation: 'src/evidence/manager.ts:261-280; close.ts:1560-1562',
+				'FINITE BY A REAL GLOBAL REAPER: one key per taskId bundle directory, and archiveEvidence enumerates the ENTIRE keyspace via listEvidenceTaskIds(directory) and deletes across it (src/evidence/manager.ts:676-738) — the maxAgeDays/maxBundles retention selects victims by comparing every remaining bundle against a global count (src/evidence/manager.ts:719-730), so unlike the ≤100-entry/≤500-KiB per-bundle cap in this row\'s `bound` field, that retention IS a keyspace bound. Belt-and-braces, "evidence" is also in ACTIVE_STATE_DIRS_TO_CLEAN and the whole tree is recursively removed at close (src/commands/close/constants.ts:253-269; src/commands/close/clean-stage.ts:176-190). CAVEAT: archiveEvidence has exactly two callers — /swarm close (src/commands/close/archive-stage.ts:531-536, 30 d / 10 bundles), where the subsequent full-tree removal makes it largely redundant, and the operator-invoked /swarm archive (src/commands/archive.ts:95, 90 d / 1000 bundles). It is not wired to any automatic interval, so mid-session reclamation requires one of those two commands to run.',
+			citation: 'src/evidence/manager.ts:261-280; src/commands/close/archive-stage.ts:498-544',
 		},
 		readBound: { pattern: 'full-file', bound: '≤500 KiB per bundle by write-side enforcement', sync: false, citation: 'src/evidence/manager.ts:275-280,387-416' },
 		lockModel: 'withEvidenceLock per bundle',
@@ -1784,7 +1784,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		legacyCompatibility: 'one-time in-place legacy migration under lock',
 		healthSignal: 'n/a',
 		owner: 'this-gate',
-		disposition: { kind: 'not-a-defect', proof: 'Hard per-bundle caps (100 entries / 500 KiB) + age/count retention + close lifecycle (src/evidence/manager.ts:261-280,665; close.ts:1560-1562).' },
+		disposition: { kind: 'not-a-defect', proof: 'Hard per-bundle caps (100 entries / 500 KiB) + age/count retention + close lifecycle (src/evidence/manager.ts:261-280,665; src/commands/close/archive-stage.ts:498-544).' },
 	},
 	{
 		id: 'phase-participation',
@@ -1836,11 +1836,11 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 			reviewedIssue: 2046,
 		},
 		writeLimits: {
-			bound: 'MAX_ROUND 10 (:18); audit-tail read ≤256 KiB; council/ dir archived+cleaned at close (close.ts:636)',
+			bound: 'MAX_ROUND 10 (:18); audit-tail read ≤256 KiB; council/ dir archived+cleaned at close (src/commands/close/constants.ts:255)',
 			scope: 'per-key',
 			keyspaceBound:
-				'FINITE BY CLOSE plus a bounded token domain: council/ is in ACTIVE_STATE_DIRS_TO_CLEAN (src/commands/close.ts:636), so every round-state/{token}.json and attempts/{token}.jsonl is removed at /swarm close on one global trigger; within a session the per-token round domain is finite (MAX_ROUND 10, src/council/council-round-state.ts:18) and the audit reader is tail-bounded (readAuditTail ≤256 KiB, src/council/council-round-state.ts:19,568).',
-			citation: 'src/council/council-round-state.ts:18-19; close.ts:636',
+				'FINITE BY CLOSE plus a bounded token domain: council/ is in ACTIVE_STATE_DIRS_TO_CLEAN (src/commands/close/constants.ts:255), so every round-state/{token}.json and attempts/{token}.jsonl is removed at /swarm close on one global trigger; within a session the per-token round domain is finite (MAX_ROUND 10, src/council/council-round-state.ts:18) and the audit reader is tail-bounded (readAuditTail ≤256 KiB, src/council/council-round-state.ts:19,568).',
+			citation: 'src/council/council-round-state.ts:18-19; src/commands/close/constants.ts:255',
 		},
 		readBound: { pattern: 'line-bounded', bound: '≤256 KiB audit tail', sync: true, citation: 'src/council/council-round-state.ts:19,568' },
 		lockModel: 'withEvidenceLock on round-state path',
@@ -1853,7 +1853,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		disposition: {
 			kind: 'retain-by-design',
 			citation:
-				'Issue #2046 durable-every-attempt semantics require the round-state/attempts authority to persist every attempted round before a verdict path can return; growth is bounded at close (council/ in ACTIVE_STATE_DIRS_TO_CLEAN, src/commands/close.ts:636), the audit reader is tail-bounded (readAuditTail ≤256 KiB, src/council/council-round-state.ts:19,568), and the per-token round domain is finite (MAX_ROUND 10, src/council/council-round-state.ts:18).',
+				'Issue #2046 durable-every-attempt semantics require the round-state/attempts authority to persist every attempted round before a verdict path can return; growth is bounded at close (council/ in ACTIVE_STATE_DIRS_TO_CLEAN, src/commands/close/constants.ts:255), the audit reader is tail-bounded (readAuditTail ≤256 KiB, src/council/council-round-state.ts:19,568), and the per-token round domain is finite (MAX_ROUND 10, src/council/council-round-state.ts:18).',
 		},
 	},
 	{
@@ -1871,8 +1871,8 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 			bound: 'one criteria file per task; council/ dir close-scoped',
 			scope: 'per-key',
 			keyspaceBound:
-				'FINITE BY REAPER, not by key domain: one key per safeId(taskId) flat file under .swarm/council/ (src/council/criteria-store.ts:39,48), and safeId only sanitizes characters for filesystem safety (:71-73) — it constrains the SHAPE of a key, never how many keys exist. The GLOBAL deleter is the close clean loop: "council" is in ACTIVE_STATE_DIRS_TO_CLEAN (src/commands/close.ts:574-583) and the loop recursively removes .swarm/council/ entire (src/commands/close.ts:1761-1773), taking every criteria file (and the sibling round-state/ and attempts/ subtrees) in one pass. CAVEAT: archive-first-gated (src/commands/close.ts:1762-1765); /swarm reset and /swarm reset-session leave council/ alone, so a session that never closes retains one file per distinct taskId passed to declare_council_criteria.',
-			citation: 'close.ts:567',
+				'FINITE BY REAPER, not by key domain: one key per safeId(taskId) flat file under .swarm/council/ (src/council/criteria-store.ts:39,48), and safeId only sanitizes characters for filesystem safety (:71-73) — it constrains the SHAPE of a key, never how many keys exist. The GLOBAL deleter is the close clean loop: "council" is in ACTIVE_STATE_DIRS_TO_CLEAN (src/commands/close/constants.ts:253-269) and the loop recursively removes .swarm/council/ entire (src/commands/close/clean-stage.ts:176-190), taking every criteria file (and the sibling round-state/ and attempts/ subtrees) in one pass. CAVEAT: archive-first-gated (src/commands/close/clean-stage.ts:176-185); /swarm reset and /swarm reset-session leave council/ alone, so a session that never closes retains one file per distinct taskId passed to declare_council_criteria.',
+			citation: 'src/commands/close/constants.ts:255',
 		},
 		readBound: { pattern: 'indexed', bound: 'single JSON per task', sync: true, citation: 'src/council/criteria-store.ts:53' },
 		lockModel: 'atomic write (single writer per task)',
@@ -1898,7 +1898,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		schemaVersion: 'EvidenceBundle-shaped entries',
 		stateClass: 'governed-content',
 		privacyClass: 'content',
-		writeLimits: { bound: 'per-phase/per-final single artifacts; evidence/ close-scoped', scope: 'session-scoped', citation: 'close.ts:568' },
+		writeLimits: { bound: 'per-phase/per-final single artifacts; evidence/ close-scoped', scope: 'session-scoped', citation: 'src/commands/close/constants.ts:256' },
 		readBound: { pattern: 'indexed', bound: 'single JSON', sync: false, citation: 'src/tools/write-final-council-evidence.ts:414' },
 		lockModel: 'withEvidenceLock; final council generation-locked',
 		crashBehavior: 'atomic writes',
@@ -1966,11 +1966,11 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 			reason: 'The spec WAL/staleness marker is the drift-gate input that must hard-block save_plan in a NEXT session with no DB handle open; single-session state, unconditionally archived+cleaned at close so the next session starts drift-free.',
 			reviewedIssue: 2036,
 		},
-		writeLimits: { bound: 'single-session drift state; spec-archive/ + spec.md + staleness + snapshot all in close clean sets', scope: 'session-scoped', citation: 'close.ts:474-503,565-574 (spec.md, spec-staleness.json, spec-snapshot.md, spec-archive)' },
+		writeLimits: { bound: 'single-session drift state; spec-archive/ + spec.md + staleness + snapshot all in close clean sets', scope: 'session-scoped', citation: 'src/commands/close/constants.ts:16-195; src/commands/close/constants.ts:253-269 (spec.md, spec-staleness.json, spec-snapshot.md, spec-archive)' },
 		readBound: { pattern: 'full-file', bound: 'bounded spec reads (effective-spec.ts:11-14)', sync: true, citation: 'src/sdd/effective-spec.ts:11-14' },
 		lockModel: 'tryAcquireLock on the staleness marker + plan lock via savePlan',
 		crashBehavior: 'WAL atomic; audit append verified by readback',
-		closePolicy: 'archived+cleaned — unconditional removal so next session starts drift-free (close.ts:460-472 docblock)',
+		closePolicy: 'archived+cleaned — unconditional removal so next session starts drift-free (src/commands/close/constants.ts:67-78 docblock)',
 		closeArrayMembership: {
 			'spec.md': 'archive+clean',
 			'spec-staleness.json': 'archive+clean',
@@ -1980,7 +1980,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		legacyCompatibility: 'n/a',
 		healthSignal: 'SPEC_DRIFT_BLOCK gate',
 		owner: 'this-gate',
-		disposition: { kind: 'not-a-defect', proof: 'Single-session drift state with atomic WAL, verified appends, and unconditional close cleanup (close.ts:460-472).' },
+		disposition: { kind: 'not-a-defect', proof: 'Single-session drift state with atomic WAL, verified appends, and unconditional close cleanup (src/commands/close/constants.ts:67-78).' },
 	},
 	{
 		id: 'workflow-wal-dirs',
@@ -2000,7 +2000,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 			reason: 'Crash-recovery WALs (coder-settlement/task-repair/task-terminals/spec-archive) must be readable before and independently of swarm.db during recovery — recovery is the one path that cannot depend on the DB being healthy; terminal states and the close-cleaned directories bound them.',
 			reviewedIssue: 2036,
 		},
-		writeLimits: { bound: 'per-task WAL files; all four dirs in ACTIVE_STATE_DIRS_TO_CLEAN', scope: 'session-scoped', citation: 'close.ts:565-574; coder-settlement.ts:54' },
+		writeLimits: { bound: 'per-task WAL files; all four dirs in ACTIVE_STATE_DIRS_TO_CLEAN', scope: 'session-scoped', citation: 'src/commands/close/constants.ts:253-269; coder-settlement.ts:54' },
 		readBound: { pattern: 'indexed', bound: 'single JSON per task', sync: true, citation: 'src/workflow/coder-settlement.ts:72' },
 		lockModel: 'per-task evidence locks; task-repair under plan lock',
 		crashBehavior: 'atomic temp+rename; terminal states survive crash; recovery resumes from WAL',
@@ -2106,7 +2106,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		readBound: { pattern: 'full-file', bound: 'bounded transitively by the configured entry cap', sync: true, citation: 'src/hooks/knowledge-store.ts:139-165,880' },
 		lockModel: 'single directory lock serializing all .swarm JSONL knowledge writes (documented trade-off :678-683)',
 		crashBehavior: 'atomic rewrites (temp+rename); append path fsync-free but single-line',
-		closePolicy: 'archived-only — ARCHIVE_ARTIFACTS (close.ts:381,452-454), deliberately NOT cleaned (cross-session knowledge)',
+		closePolicy: 'archived-only — ARCHIVE_ARTIFACTS (src/commands/close/constants.ts:28-34), deliberately NOT cleaned (cross-session knowledge)',
 		closeArrayMembership: {
 			'knowledge.jsonl': 'archive-only',
 		},
@@ -2114,7 +2114,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		legacyCompatibility: 'hive/link resolution via link.json; cohort stores resolved through identity',
 		healthSignal: 'cap eviction + archive sweeps',
 		owner: 'this-gate',
-		disposition: { kind: 'not-a-defect', proof: 'Configured entry cap with priority-aware eviction, in-place archival, and deliberate cross-session retention (src/hooks/knowledge-store.ts:880-1010; close.ts:454-456).' },
+		disposition: { kind: 'not-a-defect', proof: 'Configured entry cap with priority-aware eviction, in-place archival, and deliberate cross-session retention (src/hooks/knowledge-store.ts:880-1010; src/commands/close/constants.ts:28-34).' },
 	},
 	{
 		id: 'knowledge-events',
@@ -2246,7 +2246,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		readBound: { pattern: 'full-file', bound: '≤ cap per list (20/100/200/2000)', sync: true, citation: 'citations above' },
 		lockModel: 'directory proper-lockfile (retries 5, stale 5000) — same knowledge lock domain',
 		crashBehavior: 'atomic store rewrites under lock; quarantine moves are single-lock two-file operations',
-		closePolicy: 'knowledge-rejected.jsonl archived+cleaned (ACTIVE_STATE_TO_CLEAN close.ts:483); others untouched (bounded)',
+		closePolicy: 'knowledge-rejected.jsonl archived+cleaned (ACTIVE_STATE_TO_CLEAN src/commands/close/constants.ts:160); others untouched (bounded)',
 		closeArrayMembership: {
 			'knowledge-rejected.jsonl': 'archive+clean',
 		},
@@ -2482,7 +2482,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		crashBehavior:
 			'a crash between the JSON rename and the index sync leaves the persisted stamp mismatched against a live stat of repo-graph.json; every reader falls back to the JSON path (openFreshIndex freshness check, indexed-storage.ts:685-745) and the next successful saveGraph repairs the index in one transaction — nothing is lost because repo-graph.json stays authoritative (indexed-storage.ts:4-19)',
 		closePolicy:
-			'archived+cleaned — same treatment as swarm.db: archived via archiveSqliteSnapshot (VACUUM INTO, close.ts:1394-1408), with closeRepoMemory releasing the cached connection first to avoid Windows EBUSY on unlink (close.ts:1768-1778); -wal/-shm sidecars deliberately neither archived nor cleaned (transient, recreated on next open — close.ts:416-419,526-528)',
+			'archived+cleaned — same treatment as swarm.db: archived via archiveSqliteSnapshot (VACUUM INTO, src/commands/close/archive-stage.ts:255-313), with closeRepoMemory releasing the cached connection first to avoid Windows EBUSY on unlink (src/commands/close/clean-stage.ts:118-130); -wal/-shm sidecars deliberately neither archived nor cleaned (transient, recreated on next open — src/commands/close/constants.ts:55-64; src/commands/close/constants.ts:180-186)',
 		closeArrayMembership: {
 			'repo-memory.sqlite': 'archive+clean',
 		},
@@ -2619,14 +2619,14 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		stateClass: 'operational',
 		privacyClass: 'metadata',
 		writeLimits: {
-			bound: 'per-run memory.jsonl capped at MAX_RUN_LOG_ENTRIES 2000 FIFO via appendCappedJsonl (src/memory/run-log.ts:51,70-72); unitid-probe.jsonl capped at MAX_UNITID_PROBE_ENTRIES 2000 (src/memory/injector.ts:517,545-547); runs/ dir archived+cleaned at close (close.ts:648) and age-pruned at 30 d by the retention sweep (src/retention/sweep.ts:99)',
+			bound: 'per-run memory.jsonl capped at MAX_RUN_LOG_ENTRIES 2000 FIFO via appendCappedJsonl (src/memory/run-log.ts:51,70-72); unitid-probe.jsonl capped at MAX_UNITID_PROBE_ENTRIES 2000 (src/memory/injector.ts:517,545-547); runs/ dir archived+cleaned at close (src/commands/close/constants.ts:267) and age-pruned at 30 d by the retention sweep (src/retention/sweep.ts:99)',
 			scope: 'global',
-			citation: 'src/memory/run-log.ts:51; src/memory/injector.ts:517; src/commands/close.ts:648',
+			citation: 'src/memory/run-log.ts:51; src/memory/injector.ts:517; src/commands/close/constants.ts:267',
 		},
 		readBound: { pattern: 'full-file', bound: 'per-run reads bounded transitively by the 2000-entry cap', sync: true, citation: 'src/memory/run-log.ts:53' },
 		lockModel: 'none',
 		crashBehavior: 'append + crash-atomic FIFO compaction (temp+rename)',
-		closePolicy: 'runs/ dir archived+cleaned (ACTIVE_STATE_DIRS_TO_CLEAN close.ts:648) + 30 d sweep (src/retention/sweep.ts:99); unitid-probe is env-gated diagnostics',
+		closePolicy: 'runs/ dir archived+cleaned (ACTIVE_STATE_DIRS_TO_CLEAN src/commands/close/constants.ts:267) + 30 d sweep (src/retention/sweep.ts:99); unitid-probe is env-gated diagnostics',
 		resetPolicy: 'not reset',
 		legacyCompatibility: 'n/a',
 		healthSignal: 'n/a',
@@ -2634,7 +2634,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		disposition: {
 			kind: 'not-a-defect',
 			proof:
-				'Every run file is FIFO-capped at MAX_RUN_LOG_ENTRIES 2000 (src/memory/run-log.ts:51,70-72), the env-gated probe at MAX_UNITID_PROBE_ENTRIES 2000 (src/memory/injector.ts:517,545-547), and the runs/ directory now has a close lifecycle (ACTIVE_STATE_DIRS_TO_CLEAN, src/commands/close.ts:648) plus the 30 d sweep family (src/retention/sweep.ts:99) — the #2309 accumulation gap is closed by #2483.',
+				'Every run file is FIFO-capped at MAX_RUN_LOG_ENTRIES 2000 (src/memory/run-log.ts:51,70-72), the env-gated probe at MAX_UNITID_PROBE_ENTRIES 2000 (src/memory/injector.ts:517,545-547), and the runs/ directory now has a close lifecycle (ACTIVE_STATE_DIRS_TO_CLEAN, src/commands/close/constants.ts:267) plus the 30 d sweep family (src/retention/sweep.ts:99) — the #2309 accumulation gap is closed by #2483.',
 		},
 	},
 	{
@@ -2670,7 +2670,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		schemaVersion: 'none',
 		stateClass: 'operational',
 		privacyClass: 'metadata',
-		writeLimits: { bound: 'no per-file cap; bounded by session — archived+cleaned at close (ACTIVE_STATE_TO_CLEAN close.ts:486)', scope: 'session-scoped', citation: 'close.ts:391,486 ("Plan-scoped per-attempt outcomes — archived first, then cleaned")' },
+		writeLimits: { bound: 'no per-file cap; bounded by session — archived+cleaned at close (ACTIVE_STATE_TO_CLEAN src/commands/close/constants.ts:163)', scope: 'session-scoped', citation: 'src/commands/close/constants.ts:35-38; src/commands/close/constants.ts:161-163 ("Plan-scoped per-attempt outcomes — archived first, then cleaned")' },
 		readBound: { pattern: 'full-file', bound: 'session-scoped (close-cleaned)', sync: true, citation: 'src/services/run-memory.ts:114-140' },
 		lockModel: 'none',
 		crashBehavior: 'append; parse failures skipped',
@@ -2682,7 +2682,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		legacyCompatibility: 'n/a',
 		healthSignal: 'n/a',
 		owner: 'this-gate',
-		disposition: { kind: 'not-a-defect', proof: 'Session-scoped stream with explicit close archive+clean semantics documented in the clean list itself (close.ts:483-485).' },
+		disposition: { kind: 'not-a-defect', proof: 'Session-scoped stream with explicit close archive+clean semantics documented in the clean list itself (src/commands/close/constants.ts:160-163).' },
 	},
 	{
 		id: 'documents-cache',
@@ -2692,7 +2692,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		writerModules: ['src/evidence/documents.ts', 'src/evidence/documents-retention.ts', 'src/commands/archive.ts'],
 		writerCitations: [
 			'src/evidence/documents.ts:48 writeEvidenceDocuments — appendFile, per-record text ≤4000 chars (:9,122-136)',
-			'src/evidence/documents-retention.ts:377 pruneEvidenceDocuments — streamed line reader ≤100 MiB (:66), atomic rewrite, invoked by /swarm archive + close (:8-9, close.ts:1579-1584)',
+			'src/evidence/documents-retention.ts:377 pruneEvidenceDocuments — streamed line reader ≤100 MiB (:66), atomic rewrite, invoked by /swarm archive + close (:8-9, src/commands/close/archive-stage.ts:518-534)',
 		],
 		readerCitations: ['src/evidence/documents-retention.ts:195 readCacheRows — STREAMED with 100 MiB hard cap'],
 		schemaVersion: 'document record schema',
@@ -2725,7 +2725,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		schemaVersion: 'graph schema with workspaceRoot identity validation (:360-386)',
 		stateClass: 'derived-rebuildable',
 		privacyClass: 'metadata',
-		writeLimits: { bound: 'rebuildable from source (buildImpactMap/graph builder); archived+cleaned at close (ACTIVE_STATE_TO_CLEAN close.ts:487)', scope: 'session-scoped', citation: 'close.ts:474-503; storage.ts:339-496' },
+		writeLimits: { bound: 'rebuildable from source (buildImpactMap/graph builder); archived+cleaned at close (ACTIVE_STATE_TO_CLEAN src/commands/close/constants.ts:164)', scope: 'session-scoped', citation: 'src/commands/close/constants.ts:16-195; storage.ts:339-496' },
 		readBound: { pattern: 'full-file', bound: 'cached + validated; reflection reader hard-bounded 16 MiB', sync: true, citation: 'src/memory/reflection-service.ts:42-43,375-395' },
 		lockModel: 'no lockfile — atomic rename + mtime cache invalidation',
 		crashBehavior: 'atomic writes; corrupt graph rejected and rebuilt',
@@ -2753,11 +2753,11 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		schemaVersion: 'fingerprint schema',
 		stateClass: 'derived-rebuildable',
 		privacyClass: 'metadata',
-		writeLimits: { bound: 'bounded read (24 MiB); the sidecar is archived+cleaned at close with its sibling repo-graph.json (#2483: close.ts:430,547)', scope: 'session-scoped', citation: 'freshness.ts:43-44; close.ts:430,547' },
+		writeLimits: { bound: 'bounded read (24 MiB); the sidecar is archived+cleaned at close with its sibling repo-graph.json (#2483: src/commands/close/constants.ts:42; src/commands/close/constants.ts:166)', scope: 'session-scoped', citation: 'freshness.ts:43-44; src/commands/close/constants.ts:42; src/commands/close/constants.ts:166' },
 		readBound: { pattern: 'indexed', bound: '≤24 MiB / ≤100,256 entries', sync: true, citation: 'src/tools/repo-graph/freshness.ts:43-44' },
 		lockModel: 'none (atomic rename + probe cache TTL 30 s)',
 		crashBehavior: 'atomic',
-		closePolicy: 'archived+cleaned with its sibling repo-graph.json (#2483: ARCHIVE_ARTIFACTS close.ts:430, ACTIVE_STATE_TO_CLEAN close.ts:547)',
+		closePolicy: 'archived+cleaned with its sibling repo-graph.json (#2483: ARCHIVE_ARTIFACTS src/commands/close/constants.ts:42, ACTIVE_STATE_TO_CLEAN src/commands/close/constants.ts:166)',
 		closeArrayMembership: {
 			'repo-graph.fingerprint.json': 'archive+clean',
 		},
@@ -2768,7 +2768,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		disposition: {
 			kind: 'not-a-defect',
 			proof:
-				'The close orphan is closed: repo-graph.fingerprint.json is now in BOTH ARCHIVE_ARTIFACTS (close.ts:430) and ACTIVE_STATE_TO_CLEAN (close.ts:547), riding with its sibling repo-graph.json; reads stay bounded ≤24 MiB / ≤100,256 entries (src/tools/repo-graph/freshness.ts:43-44).',
+				'The close orphan is closed: repo-graph.fingerprint.json is now in BOTH ARCHIVE_ARTIFACTS (src/commands/close/constants.ts:42) and ACTIVE_STATE_TO_CLEAN (src/commands/close/constants.ts:166), riding with its sibling repo-graph.json; reads stay bounded ≤24 MiB / ≤100,256 entries (src/tools/repo-graph/freshness.ts:43-44).',
 		},
 	},
 	{
@@ -2834,10 +2834,10 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		category: 8,
 		pathGrammar: '.swarm/archive/swarm-{timestamp}-{suffix}/',
 		canonicalRoot: 'project-swarm',
-		writerModules: ['src/commands/close.ts'],
-		writerCitations: ['src/commands/close.ts:1256 runArchiveStage — bundle swarm-{ts}-{suffix} (:677,1262,1524); archive-first guard (:1637-1671)'],
+		writerModules: ['src/commands/close/archive-stage.ts'],
+		writerCitations: ['src/commands/close/archive-stage.ts runArchiveStage — bundle swarm-{ts}-{suffix}; archive-first guard'],
 		readerCitations: [
-			'close.ts:2234 finalize idempotency — readdir + startsWith(swarm-) (filename-only)',
+			'src/commands/close/orchestrator.ts:142-146 finalize idempotency — readdir + startsWith(swarm-) (filename-only)',
 			'session-reflection.ts:424 — filename-only scan for reflection signals',
 			'no production reader of bundle CONTENTS (verified)',
 		],
@@ -2847,10 +2847,10 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		writeLimits: {
 			bound: 'one bundle per finalize; each bundle bounded by the per-session caps of its contents; NO count/age prune across bundles',
 			scope: 'per-trigger',
-			citation: 'close.ts:1256-1613 (no prune path — verified against source)',
+			citation: 'src/commands/close/archive-stage.ts:143-497 (no prune path — verified against source)',
 		},
-		readBound: { pattern: 'directory-scan', bound: 'filename-only scans; contents never re-read', sync: false, citation: 'close.ts:2233-2235; session-reflection.ts:424' },
-		lockModel: 'finalize.lock cross-process (close.ts:2600-2613)',
+		readBound: { pattern: 'directory-scan', bound: 'filename-only scans; contents never re-read', sync: false, citation: 'src/commands/close/orchestrator.ts:142-146; session-reflection.ts:424' },
+		lockModel: 'finalize.lock cross-process (src/commands/close/orchestrator.ts:535-557)',
 		crashBehavior: 'archive-before-clean guard: active files unlinked only if archived (:1637-1671); archiveStageFailed prevents truthful-looking empty results',
 		closePolicy: 'IS the close archive',
 		resetPolicy: 'not reset; operator-managed',
@@ -2984,7 +2984,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		schemaVersion: 'doctor result schema',
 		stateClass: 'operational',
 		privacyClass: 'metadata',
-		writeLimits: { bound: 'single rewritten artifact + timestamped backups deleted by close (close.ts:1754-1762)', scope: 'global', citation: 'src/commands/close.ts:1754-1762' },
+		writeLimits: { bound: 'single rewritten artifact + timestamped backups deleted by close (src/commands/close/clean-stage.ts:192-217)', scope: 'global', citation: 'src/commands/close/clean-stage.ts:192-217' },
 		readBound: { pattern: 'indexed', bound: 'single JSON', sync: true, citation: 'src/services/config-doctor.ts:2194-2248' },
 		lockModel: 'atomic writes with Windows retry',
 		crashBehavior: 'fail-open write/read',
@@ -2996,7 +2996,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		legacyCompatibility: 'legacy numeric-hash restore support (:811-821)',
 		healthSignal: 'doctor checks themselves',
 		owner: 'this-gate',
-		disposition: { kind: 'not-a-defect', proof: 'Rewritten-per-run artifact; backups batch-deleted at close (config-doctor.ts:737; close.ts:1754-1762).' },
+		disposition: { kind: 'not-a-defect', proof: 'Rewritten-per-run artifact; backups batch-deleted at close (config-doctor.ts:737; src/commands/close/clean-stage.ts:192-217).' },
 	},
 	{
 		id: 'session-state-snapshot',
@@ -3018,7 +3018,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 			reason: 'Per-session snapshot authority (versions 1-3, incompatible quarantined) under the close/reset-cleaned session/ tree; the snapshot must load before any DB-backed service warms up.',
 			reviewedIssue: 2036,
 		},
-		writeLimits: { bound: 'single snapshot per session; session/ dir archived+cleaned at close; session-start.jsonl appends bounded by session length', scope: 'session-scoped', citation: 'close.ts:569; reset-session.ts:91-133' },
+		writeLimits: { bound: 'single snapshot per session; session/ dir archived+cleaned at close; session-start.jsonl appends bounded by session length', scope: 'session-scoped', citation: 'src/commands/close/constants.ts:257; reset-session.ts:91-133' },
 		readBound: { pattern: 'full-file', bound: 'session-scoped files', sync: false, citation: 'src/session/snapshot-reader.ts:320-347' },
 		lockModel: 'SQLite BEGIN IMMEDIATE across processes plus in-flight projection serialization',
 		crashBehavior: 'SQLite WAL recovery; strict one-time legacy import fails closed; projection/archive repair is replayable',
@@ -3027,7 +3027,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		legacyCompatibility: 'snapshot version migration 1→3',
 		healthSignal: 'quarantine presence',
 		owner: 'this-gate',
-		disposition: { kind: 'not-a-defect', proof: 'Session-scoped snapshot family under the close/reset-cleaned session/ directory with version-quarantine safety (close.ts:569; snapshot-reader.ts:334-347).' },
+		disposition: { kind: 'not-a-defect', proof: 'Session-scoped snapshot family under the close/reset-cleaned session/ directory with version-quarantine safety (src/commands/close/constants.ts:257; snapshot-reader.ts:334-347).' },
 	},
 	{
 		id: 'full-auto-state',
@@ -3309,7 +3309,8 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		pathGrammar: '.swarm/{close-summary.md, context.md, session-reflection.md, handoff.md, handoff-prompt.md, handoff-consumed.md, escalation-report.md, dark-matter.md, doc-manifest.json}',
 		canonicalRoot: 'project-swarm',
 		writerModules: [
-			'src/commands/close.ts',
+			'src/commands/close/clean-stage.ts',
+			'src/commands/close/orchestrator.ts',
 			'src/services/session-reflection.ts',
 			'src/commands/handoff.ts',
 			'src/hooks/agent-activity.ts',
@@ -3319,7 +3320,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 			'src/hooks/skill-propagation-gate.ts',
 		],
 		writerCitations: [
-			'close.ts:2465 close-summary (written AFTER clean, survives into next session); :1918-1939 context.md reset stub',
+			'src/commands/close/orchestrator.ts close-summary (written AFTER clean, survives into next session); src/commands/close/clean-stage.ts context.md reset stub',
 			'session-reflection.ts:1299 writeSessionReflection — fs.writeFile',
 			'handoff.ts:53/63 — atomic handoff briefs',
 			'agent-activity.ts:147 doFlush — context.md section rewrite (flush ≥20 events :107)',
@@ -3328,7 +3329,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		schemaVersion: 'markdown/json session documents',
 		stateClass: 'governed-content',
 		privacyClass: 'content',
-		writeLimits: { bound: 'single rewritten session documents (close-summary/handoff atomic; context.md sectioned)', scope: 'session-scoped', citation: 'close.ts:474-503 (all but close-summary archived+cleaned)' },
+		writeLimits: { bound: 'single rewritten session documents (close-summary/handoff atomic; context.md sectioned)', scope: 'session-scoped', citation: 'src/commands/close/constants.ts:16-195 (all but close-summary archived+cleaned)' },
 		readBound: { pattern: 'full-file', bound: 'single documents', sync: false, citation: 'handoff.ts:45-46 consumers' },
 		lockModel: 'atomic writes; snapshot in-flight serialization',
 		crashBehavior: 'atomic rewrites; failures logged',
@@ -3348,7 +3349,7 @@ export const RETENTION_REGISTRY: readonly RetentionRow[] = [
 		legacyCompatibility: 'n/a',
 		healthSignal: 'n/a',
 		owner: 'this-gate',
-		disposition: { kind: 'not-a-defect', proof: 'Single-session documents with explicit archive/clean/stub semantics in the close lists themselves (close.ts:474-503,1918-1939,2465).' },
+		disposition: { kind: 'not-a-defect', proof: 'Single-session documents with explicit archive/clean/stub semantics in the close lists themselves (src/commands/close/constants.ts:16-95; src/commands/close/clean-stage.ts:364-380; src/commands/close/orchestrator.ts:397-406).' },
 	},
 	{
 		id: 'command-reports',
